- implemented usage of RabbitMQConfiguration.DefaultPrefetchCount
- https://github.com/Shuttle/shuttle-esb/issues/11

+++ source/Shuttle.ESB.RabbitMQ/RabbitMQConfiguration.cs
@@ -30,7 +30,7 @@ namespace Shuttle.ESB.RabbitMQ
 		public int RemoteQueueTimeoutMilliseconds { get; set; }
 		public int ConnectionCloseTimeoutMilliseconds { get; set; }
 		public int OperationRetryCount { get; set; }
-		public int DefaultPrefetchCount { get; set; }
+		public ushort DefaultPrefetchCount { get; set; }
 
 		public static RabbitMQConfiguration Default()
 		{
